test(refactor): share artifact fixtures and refresh import snapshots
